refactor(gui-app): lift the browsers panel's row layer out of its components
The row reconciliation, the disambiguating labels, the search predicate, the
close-and-retire-the-tile action and the driven-by coalescing were all private
to the desktop Browsers panel's two component files. A second surface listing
the same tabs would have had to restate every one of them, and a restated rule
is a rule that drifts.

Each moves to a module of its own with no behavior change:

- `use-browser-tab-rows.ts` - stable row order, settled identities, the
  hostname/tab-id disambiguation and the title-or-URL filter.
- `use-browser-tab-close.ts` - ask the host, then close the tile the row points
  at, folding the host's own `closing` status in so an agent's close reads the
  same as ours. The close button's label comes with it.
- `use-coalesced-browser-tab-drivers.ts` - the bidirectional delay behind the
  driven-by glyph, moved verbatim; it lives beside the React-free timing rule
  it drives rather than inside a component file, which also keeps that file
  exporting components only.

`browserTabDriverNames` joins the coalescing module, and the panel's four
load / unavailable / empty / no-results states are exported so a second surface
mounts the same copy instead of writing its own.
